refactor(`Layout`): 스타일 파일로 분할
fix: container import file change

chore: 리드미 제거
