--- src/views/components/templates/README.md
@@ -1,1 +1,0 @@
-# Template
