+++ src/views/containers/Layout/index.ts
@@ -1,4 +1,4 @@
-import Layout from '@template/layout';
+import Layout from '@template/Layout';
 import { connect } from 'react-redux';
 
 const mapReduxStateToReactProps = ({ themeState }: any) => {
